requirements.txt
requests~=2.32.4
PyYAML~=6.0.2
python-dotenv~=1.1.0
playwright~=1.52.0